Move some functions into methods of `FSEnv`
