+++ internal/app/forklift/versioning-repos.go
@@ -3,12 +3,10 @@ package forklift
 import (
 	"fmt"
 	"io/fs"
-	"path/filepath"
 	"strings"
 	"time"
 
 	"github.com/blang/semver/v4"
-	"github.com/bmatcuk/doublestar/v4"
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v3"
 
@@ -64,6 +62,22 @@ func CompareVersionedRepos(r, s VersionedRepo) int {
 }
 
 // RepoVersionConfig
+
+// loadRepoVersionConfig loads a RepoVersionConfig from a specified file path in the provided base
+// filesystem.
+func loadRepoVersionConfig(fsys pallets.PathedFS, filePath string) (RepoVersionConfig, error) {
+	bytes, err := fs.ReadFile(fsys, filePath)
+	if err != nil {
+		return RepoVersionConfig{}, errors.Wrapf(
+			err, "couldn't read repository version config file %s/%s", fsys.Path(), filePath,
+		)
+	}
+	config := RepoVersionConfig{}
+	if err = yaml.Unmarshal(bytes, &config); err != nil {
+		return RepoVersionConfig{}, errors.Wrap(err, "couldn't parse repository version config")
+	}
+	return config, nil
+}
 
 func (l RepoVersionConfig) IsCommitLocked() bool {
 	return l.Commit != ""
@@ -130,107 +144,4 @@ func (l RepoVersionConfig) Version() (string, error) {
 		return "", errors.Wrap(err, "couldn't determine pseudo-version")
 	}
 	return pseudoversion, nil
-}
-
-// Loading
-
-const versionedReposDirName = "repositories"
-
-// TODO: this should be a method on FSEnv
-func VersionedReposFS(envFS fs.FS) (fs.FS, error) {
-	return fs.Sub(envFS, versionedReposDirName)
-}
-
-func LoadVersionedRepo(reposFS fs.FS, repoPath string) (VersionedRepo, error) {
-	vcsRepoPath, repoSubdir, err := SplitRepoPathSubdir(repoPath)
-	if err != nil {
-		return VersionedRepo{}, errors.Wrapf(
-			err, "couldn't parse path of version-locked Pallet repo %s", repoPath,
-		)
-	}
-	repo := VersionedRepo{
-		VCSRepoPath: vcsRepoPath,
-		RepoSubdir:  repoSubdir,
-	}
-
-	repo.Config, err = loadRepoVersionConfig(
-		reposFS, filepath.Join(repoPath, "forklift-repo.yml"),
-	)
-	if err != nil {
-		return VersionedRepo{}, errors.Wrapf(err, "couldn't load repo version config for %s", repoPath)
-	}
-
-	return repo, nil
-}
-
-// SplitRepoPathSubdir splits paths of form github.com/user-name/git-repo-name/pallets-repo-subdir
-// into github.com/user-name/git-repo-name and pallets-repo-subdir.
-func SplitRepoPathSubdir(repoPath string) (vcsRepoPath, repoSubdir string, err error) {
-	const sep = "/"
-	pathParts := strings.Split(repoPath, sep)
-	if pathParts[0] != "github.com" {
-		return "", "", errors.Errorf(
-			"pallet repository %s does not begin with github.com, but support for non-Github "+
-				"repositories is not yet implemented",
-			repoPath,
-		)
-	}
-	const splitIndex = 3
-	if len(pathParts) < splitIndex {
-		return "", "", errors.Errorf(
-			"pallet repository %s does not appear to be within a Github Git repository", repoPath,
-		)
-	}
-	return strings.Join(pathParts[0:splitIndex], sep), strings.Join(pathParts[splitIndex:], sep), nil
-}
-
-func loadRepoVersionConfig(reposFS fs.FS, filePath string) (RepoVersionConfig, error) {
-	bytes, err := fs.ReadFile(reposFS, filePath)
-	if err != nil {
-		return RepoVersionConfig{}, errors.Wrapf(
-			err, "couldn't read repo version config file %s", filePath,
-		)
-	}
-	config := RepoVersionConfig{}
-	if err = yaml.Unmarshal(bytes, &config); err != nil {
-		return RepoVersionConfig{}, errors.Wrap(err, "couldn't parse repo version config")
-	}
-	return config, nil
-}
-
-// Listing
-
-func ListVersionedRepos(envFS fs.FS) ([]VersionedRepo, error) {
-	reposFS, err := VersionedReposFS(envFS)
-	if err != nil {
-		return nil, errors.Wrap(
-			err, "couldn't open directory for Pallet repositories in local environment",
-		)
-	}
-	files, err := doublestar.Glob(reposFS, "**/forklift-repo.yml")
-	if err != nil {
-		return nil, errors.Wrap(err, "couldn't search for Pallet repo versioning configs")
-	}
-
-	repoPaths := make([]string, 0, len(files))
-	repoMap := make(map[string]VersionedRepo)
-	for _, filePath := range files {
-		repoPath := filepath.Dir(filePath)
-		if _, ok := repoMap[repoPath]; ok {
-			return nil, errors.Errorf(
-				"versioned repository %s repeatedly defined in the local environment", repoPath,
-			)
-		}
-		repoPaths = append(repoPaths, repoPath)
-		repoMap[repoPath], err = LoadVersionedRepo(reposFS, repoPath)
-		if err != nil {
-			return nil, errors.Wrapf(err, "couldn't load versioned repo from %s", repoPath)
-		}
-	}
-
-	orderedRepos := make([]VersionedRepo, 0, len(repoPaths))
-	for _, repoPath := range repoPaths {
-		orderedRepos = append(orderedRepos, repoMap[repoPath])
-	}
-	return orderedRepos, nil
 }
